added more unit test coverage for EditionExists
